requirements.txt
pyyaml
flask
requests
xmljson
thrift
boto3
pytz
aliyun-python-sdk-core
aliyun-python-sdk-slb
httplib2
aliyun-python-sdk-cms
aliyun-python-sdk-vpc
oss2
six
ks3sdk